+++ src/cli/import.ts
@@ -49,17 +49,86 @@ const datasets = blzFile
 // ignore the rest.
 const masterDatasets = datasets.filter((dataset) => dataset.master);
 
+interface CheckDigitData {
+  [key: string]: number[];
+}
+
 // Write all checkDigit methods with assigned BLZs
 // If save over 20kb if use the checkDigit method as index and put the BLZs
 // as values instead of the other way round.
-const dataCheckDigit: { [key: string]: number[] } = {};
+const dataCheckDigit: CheckDigitData = {};
 masterDatasets.forEach((dataset) => {
   if (typeof dataCheckDigit[dataset.checkDigitMethod] === "undefined") {
     dataCheckDigit[dataset.checkDigitMethod] = [];
   }
   dataCheckDigit[dataset.checkDigitMethod].push(dataset.blz);
 });
+
+// If file doesn't exist just write it and we're done
+if (!fs.existsSync(`${__dirname}/../data/current.json`)) {
+  fs.writeFileSync(
+    `${__dirname}/../data/current.json`,
+    JSON.stringify(dataCheckDigit)
+  );
+
+  process.exit(0);
+}
+
+if (
+  typeof process.argv[2] !== "string" ||
+  !process.argv[2].match(
+    /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:UTC|[+-]\d{4})/
+  )
+) {
+  console.error("Please provide a valid-from datetime");
+  process.exit(1);
+}
+
+// Get the current/previous data
+const currentFile = fs.readFileSync(`${__dirname}/../data/current.json`);
+const current = JSON.parse(currentFile.toString()) as CheckDigitData;
+
+const currentMethods = Object.keys(current).sort();
+const nextMethods = Object.keys(dataCheckDigit).sort();
+
+// Exit if methods have changed because this may invoke code adjustments
+if (JSON.stringify(currentMethods) !== JSON.stringify(nextMethods)) {
+  console.warn("Check digits methods have changed!");
+  process.exit(2);
+}
+
+// Check which BLZ have been added and which removed
+const dataAddedCheckDigit: CheckDigitData = {};
+const dataRemovedCheckDigit: CheckDigitData = {};
+for (const method of currentMethods) {
+  const currentMethodBLZs = current[method];
+  const nextMethodBLZs = dataCheckDigit[method];
+
+  for (const currentMethodBLZ of currentMethodBLZs) {
+    if (!nextMethodBLZs.includes(currentMethodBLZ)) {
+      if (typeof dataRemovedCheckDigit[method] === "undefined") {
+        dataRemovedCheckDigit[method] = [];
+      }
+      dataRemovedCheckDigit[method].push(currentMethodBLZ);
+    }
+  }
+
+  for (const nextMethodBLZ of nextMethodBLZs) {
+    if (!currentMethodBLZs.includes(nextMethodBLZ)) {
+      if (typeof dataAddedCheckDigit[method] === "undefined") {
+        dataAddedCheckDigit[method] = [];
+      }
+      dataAddedCheckDigit[method].push(nextMethodBLZ);
+    }
+  }
+}
+
+// Write the diff data to next.json
 fs.writeFileSync(
-  `${__dirname}/../data/current.json`,
-  JSON.stringify(dataCheckDigit)
+  `${__dirname}/../data/next.json`,
+  JSON.stringify({
+    add: dataAddedCheckDigit,
+    remove: dataRemovedCheckDigit,
+    valid: process.argv[2],
+  })
 );