+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2022-11-13 [v1.2201.0]
+
+* Additionally includes data valid from 2022-12-05
+
 ## 2022-10-31 [v1.2200.6]
 
 * Fixed methods C0 ; [#7](https://github.com/baumerdev/ibantools-germany/pull/7)